requirements.txt
Django==5.1
psycopg-pool==3.2.2
djangorestframework==3.15.2
psycopg==3.2.1
psycopg-binary==3.2.1
drf-spectacular==0.27.2
pillow==10.4.0